requirements.txt
tensorflow==2.7
sklearn
tqdm
pandas
biopython
pillow
matplotlib
snakemake
pyyaml